requirements.txt
absl_py==1.3.0
Flask==1.1.2
Flask==2.2.3
Flask_WTF==1.1.1
matplotlib==3.5.2
matplotlib==3.5.3
matplotlib==3.6.2
numpy==1.21.5
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.5.4.60
paddleocr==2.6.1.3
paddleocr.egg==info
Requests==2.30.0
Werkzeug==2.0.3
Werkzeug==2.2.2
Werkzeug==2.2.3
WTForms==3.0.1
xmltodict==0.13.0
